requirements.txt
discord.py
pandas
openpyxl
asyncio